Split loading and indexing reports into workers
Using worker processes lets us get around garbage collection issues with
node. The worker-farm module handles starting worker processes,
assigning jobs, and retrying when a worker process fails.
