+++ package.json
@@ -42,7 +42,8 @@
     "minimist": "^1.2.0",
     "numeral": "^1.5.3",
     "wintersmith": "^2.3.1",
-    "wintersmith-ejs": "^0.1.4"
+    "wintersmith-ejs": "^0.1.4",
+    "worker-farm": "^1.3.1"
   },
   "devDependencies": {
     "intercept-stdout": "^0.1.2",
